解决当 apart_level 的日志文件数量大于 max_files 时，日志文件不但不会生成，还有可能会被莫名删除的问题

+++ src/think/log/driver/File.php
@@ -138,7 +138,7 @@ class File implements LogHandlerInterface
             $files = glob($this->config['path'] . '*.log');
 
             try {
-                if (count($files) > $this->config['max_files']) {
+                if (count($files) > $this->config['max_files'] && count($files) > count($this->config['apart_level'])) {
                     set_error_handler(function ($errno, $errstr, $errfile, $errline) {});
                     unlink($files[0]);
                     restore_error_handler();
